Add UserCurvePattern
While playing around with various patterns, I noticed there was no good way to generate e.g. a few moments at high intensity which then tapers off. My proposed solution is this UserCurve pattern, which interpolates between an arbitrary set of points. I implemented this using a canonical spline; the math is actually super interesting, but I don't think this is the place for it -- check Wikipedia if you're interested. :)

I played around with this a fair bit using a plot preview, and also checked the behavior on real haptics, and it seems fine. Note that particularly sharp turns and various tension values can look kinda jank; I clamped the output to 0..1 (although if you check the combiner work maybe I shouldn't?) to keep the curve from trying to go to infinity or something.