Add a pattern preview plot
While I was working on the UserCurvePattern, I realized it would be pretty challenging to validate the spline interpolation without something like this. Then I saw it was an open issue on the repo, and, well, I might as well, right?

This does a pretty simple ImGui.PlotLines by generating a temporary Pattern instance with the working trigger/pattern config. It's a bit stupid, but produces a decent sample of how the device intensity will change over the course of the trigger execution.

It's worth noting that this could be very inefficient and/or hard on a user's CPU. A smarter implementation would, like, find discontinuities and inflection points and only use those points and subdivisions of continuous regions to generate a small number of points for ImGui.PlotLines... or maybe would just preserve the graph and only update it on changes. I dunno!

Resolves(?) issue #18