Pattern preview: Switch to nicer ImPlot
I decided I was dissatisfied iwth the simple ImGui.PlotLines approach; I wanted a fixed Y axis from 0 to 1. Turns out, Dalamud provides ImPlot for such applications. So I rewrote the above, and shuffled the preview button back to the top next to the pattern type combo box. This generates a much prettier plot, with a clear/stable indication of intensity relative to other patterns.

In the process of working on these, I noticed that square and random patterns both seem to be independent of trigger/expiry time. This makes them pretty unpredictable -- I'd regard these as issues/bugs, personally. I may propose PRs to "fix" them, by affixing the square pattern to the trigger time and seeding the random pattern with the trigger time. (Random would still be really trippy with the preview, though. Would have to also modify it to only rerender the pattern when a parameter changes or something.)