Discover pattern implementations at runtime
Rather than hardcoding the list of pattern types in numerous places, we should have a single list of all known pattern types. And if that list is dynamically generated from all implementations of an interface which can provide the various functions needed by PatternFactory and some extra metadata to boot, all the better. In my mind, this is a far cleaner approach to the internal code.

This is meant to significantly ease the addition of new patterns. I have in mind a similar refactor for trigger types.

I realize that bringing ImGui into the pattern sources might be seen as a breach of containment/separation of concerns, where previously the pattern files were strictly concerned with modelling the configuration and providing the intensity function, but now must also define their UI. While we could make a declarative UI for patterns, I think the nature of immediate-mode user interfaces kind of demands this integration. *shrug*