Add various combination modes (#26)
"Average" is probably one of the oddest options we could default to, isn't it? If you have one strong signal and a low signal kicks in, average would _reduce_ the vibration intensity, even though you now have two "sources" of haptics.

So I added an enum to Configuration containing four combination modes, propagated the configuration object through to DeviceService and Device (had to reorder creation of DeviceService slightly but I don't see any conflicts), implemented the combination modes in Device, and added the UI to the advanced tab of PluginUI.

The four modes, for reference:
- Average: The original behavior, takes the mean of the input signals.
- Max: Chooses the largest input signal.
- Add: Sums all the signals. (The result is only clamped to 0..1 afterward, so if we ever have negative signals, they can "penalize" the output signal.)
- Asymptotic: 1 - (1 - sig1)*(1 - sig2)*...*(1 - sigN). Again, input signals are not pre-clamped; a negative signal can reduce the output and a greater-than-1 signal can increase it. (The final value is still clamped though.)