feat: Add a pattern preview plot (#28)
* Add a pattern preview plot

While I was working on the UserCurvePattern, I realized it would be pretty challenging to validate the spline interpolation without something like this. Then I saw it was an open issue on the repo, and, well, I might as well, right?

This does a pretty simple ImGui.PlotLines by generating a temporary Pattern instance with the working trigger/pattern config. It's a bit stupid, but produces a decent sample of how the device intensity will change over the course of the trigger execution.

It's worth noting that this could be very inefficient and/or hard on a user's CPU. A smarter implementation would, like, find discontinuities and inflection points and only use those points and subdivisions of continuous regions to generate a small number of points for ImGui.PlotLines... or maybe would just preserve the graph and only update it on changes. I dunno!

Resolves(?) issue #18

* Pattern preview: Switch to nicer ImPlot

I decided I was dissatisfied iwth the simple ImGui.PlotLines approach; I wanted a fixed Y axis from 0 to 1. Turns out, Dalamud provides ImPlot for such applications. So I rewrote the above, and shuffled the preview button back to the top next to the pattern type combo box. This generates a much prettier plot, with a clear/stable indication of intensity relative to other patterns.

In the process of working on these, I noticed that square and random patterns both seem to be independent of trigger/expiry time. This makes them pretty unpredictable -- I'd regard these as issues/bugs, personally. I may propose PRs to "fix" them, by affixing the square pattern to the trigger time and seeding the random pattern with the trigger time. (Random would still be really trippy with the preview, though. Would have to also modify it to only rerender the pattern when a parameter changes or something.)